+++ packages/antd-management-fast-framework/src/framework/DataModal/Base/index.jsx
@@ -335,7 +335,7 @@ class Base extends BaseWindow {
 
     return (
       <ModalExtra
-        destroyOnClose={this.destroyOnClose || false}
+        destroyOnHidden={this.destroyOnHidden || false}
         flag={this.getVisibleFlag()}
         icon={this.renderPresetTitleIcon()}
         titlePrefix={this.buildTitlePrevText()}