+++ packages/antd-management-fast-framework/src/framework/DataMultiPageView/MultiPageDrawer/index.jsx
@@ -651,7 +651,7 @@ class MultiPageDrawer extends MultiPage {
         titlePrefix={this.buildTitlePrevText()}
         title={this.buildTitleText()}
         subtitle={this.buildTitleSubText()}
-        destroyOnClose={false}
+        destroyOnHidden={false}
         className={styles.containorBox}
         width={width}
         placement="right"