+++ examples/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.9.85
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.7.53
+  - antd-management-fast-framework@2.12.59
+  - antd-management-fast-formily@1.0.148
+  - antd-management-fast-design-playground@1.0.208
+  - antd-management-fast-flow@1.0.136
+
 ## 1.9.84
 
 ### Patch Changes